# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConciergeActionOut": {
        "description": "A suggested next step the UI renders as a button — navigation to a (separately gated) page.\nThe concierge never executes mutations; this only routes the user to the relevant tool.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "volume_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Id"
          }
        },
        "required": [
          "label",
          "route"
        ],
        "title": "ConciergeActionOut",
        "type": "object"
      },
      "ConciergeAnswerOut": {
        "description": "A concierge answer: the prose, the tool the model chose, grounded citations + next steps.",
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ConciergeActionOut"
            },
            "title": "Actions",
            "type": "array"
          },
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/ConciergeCitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "considered": {
            "title": "Considered",
            "type": "integer"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "tool",
          "considered",
          "citations"
        ],
        "title": "ConciergeAnswerOut",
        "type": "object"
      },
      "ConciergeAskRequest": {
        "description": "A natural-language question for the concierge, with optional host/volume + page hints.",
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/ConciergeTurn"
            },
            "maxItems": 20,
            "title": "History",
            "type": "array"
          },
          "host_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Id"
          },
          "page": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "tool": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "volume_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Id"
          }
        },
        "required": [
          "question"
        ],
        "title": "ConciergeAskRequest",
        "type": "object"
      },
      "ConciergeCitationOut": {
        "description": "A server-built reference to a real result row the answer is grounded in (never the LLM's).",
        "properties": {
          "entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "host_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Id"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "volume_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Id"
          }
        },
        "required": [
          "label"
        ],
        "title": "ConciergeCitationOut",
        "type": "object"
      },
      "ConciergeTurn": {
        "description": "One prior turn of the conversation, supplied so follow-ups resolve (memory; ADR-035).",
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ConciergeTurn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Local username/password login (ADD 13 §1).",
        "properties": {
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MfaVerifyRequest": {
        "description": "A TOTP code for step-up / enrollment confirmation.",
        "properties": {
          "code": {
            "maxLength": 8,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "MfaVerifyRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Fathomline (built on the Fathom engine). The read/query routes are RBAC- and scope-gated; the agent ingest routes authenticate by mTLS client-cert fingerprint behind the proxy; the remediation, preview and deployment surfaces are default-OFF and individually gated. See the project docs for auth and deployment.",
    "summary": "The storage-estate analyzer — read/query surface, agent ingest, and the default-OFF write/deploy surfaces.",
    "title": "Fathomline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate a local user and mint a session cookie (audited; no enumeration).",
        "operationId": "post_login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/mfa/verify": {
      "post": {
        "description": "Verify a TOTP code and stamp step-up freshness on the session (ADD 13 §4).",
        "operationId": "post_mfa_verify_api_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Mfa Verify",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/concierge/ask": {
      "post": {
        "description": "Answer a natural-language storage question (read-only; scope-enforced).",
        "operationId": "concierge_ask_api_v1_concierge_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConciergeAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciergeAnswerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Concierge Ask",
        "tags": [
          "concierge"
        ]
      }
    },
...
  }
}